requirements.txt
flask==2.3.3
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.1
python-docx==0.8.11
pdfminer.six==20221105
numpy==1.24.3
pandas==2.0.3
werkzeug==2.3.7
reportlab==4.0.4